% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliplot.R
\name{caliplot}
\alias{caliplot}
\title{Calibrate sensitivity parameters based on pre-treatment variables}
\usage{
caliplot(x, trt, y, gamma_seq)
}
\arguments{
\item{x}{a \code{tibble} or data frame with observed pre-treatment variables}

\item{trt}{a vector with binary treatment indicators}

\item{y}{a vector with outcomes}

\item{gamma_seq}{a vector with chosen values for sensitivity parameter \eqn{\gamma_t}}
}
\description{
Calibrate the magnitude of sensitivity parameters to the amount of variation in the treatment 
             assignment \code{T} that is explained by outcome \code{Y(t)}, above and beyond what is accounted
             for by observed pre-treatment variables \code{X}.
}
\section{Details}{
 \code{caliplot} returns a plot of sensitivity parameter \eqn{\gamma_t} vs partial coefficient of 
         variation from outcome \code{Y}, \eqn{\rho^2_{Y|X}}. For comparison, the largest four partial coefficients 
         of variation from covariates are also plotted if the number of observed covariates is larger or equal to four;
         otherwise, all the partial coefficients of variation from covariates are plotted.
}

\examples{
# Observed Pre-treatment Variables 
x = NHANES \%>\% select(-one_of("trt_dbp", "ave_dbp"))

# Treatment 
trt = NHANES \%>\% select(trt_dbp)

# Outcomes 
y = NHANES \%>\% select(ave_dbp)

# Sensitivity Parameter Sequence 
gamma = seq(0.01, 0.1, by = 0.001)

# plot 
caliplot(x, trt, y, gamma)
}
