% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_ate.R
\name{contour_ate}
\alias{contour_ate}
\title{Visualize ATE Estimates by Contour Plot}
\usage{
contour_ate(x_trt, y_trt, x_ctrl, y_ctrl, 
            largest_effect, gamma_length = 11, ...)
}
\arguments{
\item{x_trt}{a \code{tibble} or data frame with observed pre-treatment variables for the treatment group}

\item{y_trt}{a vector with outcomes for the treatment group}

\item{x_ctrl}{a \code{tibble} or data frame with observed pre-treatment variables for the control group}

\item{y_ctrl}{a vector with outcomes for the control group}

\item{largest_effect}{the largest magnitude of sensitivity parameter to be considered, chosen from \code{\link{caliplot}}}

\item{gamma_length}{chosen length of sensitivity parameter sequence, which needs to be an odd integer}

\item{joint}{logical. If TURE, the mean surface and residual variance will be estimated jointly for both treatment 
groups; if FALSE (default), the mean surface and residual variance will be estimated independently for
each treatment group.}

\item{...}{arguments passed to the function \code{BART::wbart}}
}
\description{
Visualize the same level of Average Treatment Effect (ATE)
             for different choices of sensitivity parameters with contour lines.
}
\section{Details}{
 
         For analysis details, please see \code{\link{heatmap_ate}}
}

\examples{
# Observed data in treatment group 
NHANES_trt <- NHANES \%>\% dplyr::filter(trt_dbp == 1)
x_trt <- NHANES_trt \%>\% select(-one_of("trt_dbp", "ave_dbp"))
y_trt <- NHANES_trt \%>\% select(ave_dbp)

# Observed data in control group
NHANES_ctrl <- NHANES \%>\% dplyr::filter(trt_dbp == 0)
x_ctrl <- NHANES_ctrl \%>\% select(-one_of("trt_dbp", "ave_dbp"))
y_ctrl <- NHANES_ctrl \%>\% select(ave_dbp)

# ATE Contour Plot 
contour_ate(x_trt, y_trt, x_ctrl, y_ctrl, largest_effect = 0.05)
contour_ate(x_trt, y_trt, x_ctrl, y_ctrl, largest_effect = 0.05, joint = TRUE)
}
