% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribbon_qte.R
\name{ribbon_qte}
\alias{ribbon_qte}
\title{Visualize QTE Estimates by Ribbon Plot}
\usage{
ribbon_qte(x_trt, y_trt, x_ctrl, y_ctrl, gamma_select, joint = FALSE)
}
\arguments{
\item{x_trt}{a \code{tibble} or data frame with observed pre-treatment variables for the treatment group}

\item{y_trt}{a vector with outcomes for the treatment group}

\item{x_ctrl}{a \code{tibble} or data frame with observed pre-treatment variables for the control group}

\item{y_ctrl}{a vector with outcomes for the control group}

\item{gamma_select}{a matrix with selected settings for sensitivity parameters in rows}
}
\description{
Visualize the 95\% posterior credible band of the quantile treatment effects for
       selected settings of the sensitivity parameters.
}
\section{Details}{
 
         For analysis details, please see \code{\link{heatmap_qte}}
}

\examples{
# Observed data in treatment group
NHANES_trt <- NHANES \%>\% dplyr::filter(trt_dbp == 1)
x_trt <- NHANES_trt \%>\% select(-one_of("trt_dbp", "ave_dbp"))
y_trt <- NHANES_trt \%>\% select(ave_dbp)

# Observed data in control group
NHANES_ctrl <- NHANES \%>\% dplyr::filter(trt_dbp == 0)
x_ctrl <- NHANES_ctrl \%>\% select(-one_of("trt_dbp", "ave_dbp"))
y_ctrl <- NHANES_ctrl \%>\% select(ave_dbp)

# Ribbon Plot of QTE
gamma_select = rbind(c(0, 0), c(-0.03, 0.05), c(0.05, -0.02), c(-0.01, 0.01))
ribbon_qte(x_trt, y_trt, x_ctrl, y_ctrl, gamma_select)
ribbon_qte(x_trt, y_trt, x_ctrl, y_ctrl, gamma_select, joint = TRUE)
}
